//Classes Created ScheduleUtitlity a class that has a method to convert timeslots to time strings. Created ReserveActivity splitting responsibility from ScheduleActivity.
//ScheduleActivity Parts
Reworked ClassroomSpinner to properly work on user click and not when the user activates the activity. Separated RadioButton responsibility in it's own class.

//Cell
Created a Timeday Class which is used for setting tags in each cell with the time and day.
Created a background label which changes color on select.

//Login
Temporarily Disabled Login check because the Api changed and priorities were on different parts of the system